Fix LambDynamicLights integration on NeoForge.
