+++ gradle.properties
@@ -37,5 +37,5 @@ mod_license=Apache-2.0
 
 
 # Optional integrations
-# LambDynamicLights 4.x (for compileOnly usage via Modrinth Maven)
-lambdynamiclights_version=4.3.3+1.21.8
+# LambDynamicLights 4.x (for compileOnly usage)
+lambdynamiclights_version=4.5.1+1.21.8
