+++ modules/fabric/src/main/resources/fabric.mod.json
@@ -23,7 +23,7 @@
       "net.oxcodsnet.beltborne_lanterns.fabric.client.BLFabricClient"
     ],
     "lambdynlights:initializer": [
-      "net.oxcodsnet.beltborne_lanterns.fabric.compat.LDL4FabricInitializer"
+      "net.oxcodsnet.beltborne_lanterns.common.LambDynLightsInitializer"
     ],
     "modmenu": [
       "net.oxcodsnet.beltborne_lanterns.fabric.client.BLFabricModMenu"
