+++ modules/neoforge/build.gradle
@@ -40,14 +40,6 @@ dependencies {
 
     common(project(path: ':modules:common', configuration: 'namedElements')) { transitive = false }
     shadowBundle project(path: ':modules:common', configuration: 'transformProductionNeoForge')
-
-    // Optional: LambDynamicLights 4.x API for typed integration (compile-only)
-    // Use Mojang-mapped capability to ensure generics are remappable to Yarn for NeoForge
-    modCompileOnly("dev.lambdaurora.lambdynamiclights:lambdynamiclights-api:${rootProject.lambdynamiclights_version}") {
-        capabilities {
-            requireCapability("dev.lambdaurora.lambdynamiclights:api-mojmap")
-        }
-    }
 }
 
 processResources {
