+++ modules/neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -43,3 +43,6 @@ side = "BOTH"
 
 [[mixins]]
 config = "beltborne_lanterns.mixins.json"
+
+[modproperties.beltborne_lanterns."yumi:entrypoints"]
+"lambdynlights:initializer" = "net.oxcodsnet.beltborne_lanterns.common.LambDynLightsInitializer"
