--- modules/neoforge/src/main/resources/META-INF/services/dev.lambdaurora.lambdynlights.api.DynamicLightsInitializer
@@ -1,1 +1,0 @@
-net.oxcodsnet.beltborne_lanterns.neoforge.compat.LDL4NeoForgeInitializer
